requirements.txt
folium==0.17.0
haversine==2.8.1
matplotlib==3.8.4
matplotlib-inline==0.1.7
numpy==1.26.4
pandas==2.2.2
plotly==5.23.0
streamlit==1.38.0
streamlit-folium==0.22.0
Pillow==9.2.0
